Update Info.plist
With the Xcode 7.0 update, App Transport Security is enabled for simulators and devices running iOS 9. See [App Transport Security Technote](https://developer.apple.com/library/prerelease/ios/technotes/App-Transport-Security-Technote/).

For this sample we have created a temporary exception for the following domain in the plist:

- outlook.office365.com

If this exception is not included, all calls into the Office 365 API will fail in this app when deployed to an iOS 9 simulator in Xcode.